Adding and removing from lists UX niceities:
- Automatically choose a list when none is chosen, defaulting to the last updated.
- Create a new default list if none is present
- Allow selecting active list

+++ app/models/personal_space_list.rb
@@ -4,6 +4,7 @@ class PersonalSpaceList < ApplicationRecord
   belongs_to :user
 
   has_one :active_personal_space_list, dependent: :destroy
+  accepts_nested_attributes_for :active_personal_space_list
   has_and_belongs_to_many :spaces
 
   validates :title, presence: true
@@ -24,23 +25,38 @@ class PersonalSpaceList < ApplicationRecord
     spaces.exclude?(space)
   end
 
+  def active
+    active?
+  end
+
   def active?
     active_personal_space_list.present?
   end
 
+  def inactive?
+    active_personal_space_list.blank?
+  end
+
   def activate
+    # Deactivate any active lists for this user
+    ActivePersonalSpaceList.where(user:).destroy_all
     ActivePersonalSpaceList.create(user:, personal_space_list: self)
   end
 
   def deactivate
-    active_personal_space_list.destroy
+    active_personal_space_list.destroy if active?
   end
 
   # self.METHOD_NAME makes this a Class method, the others are instance methods
   # E.g it can be called like this: PersonalSpaceList.create_default_list_for(user)
   # instead of this: PersonalSpaceList.new.create_default_list_for(user)
   def self.create_default_list_for(user, active: true)
-    default_list = create(user:, title: I18n.t("personal_space_lists.default_list_name", name: user.name))
+    default_list = order(updated_at: :desc).find_or_initialize_by(user:)
+    if default_list.title.blank?
+      default_list.update(title: I18n.t("personal_space_lists.default_list_name",
+                                        name: user.name))
+    end
+    default_list.save
     default_list.activate if active
 
     default_list
